Split options into more fine-grained structs
